python/requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
clarifai==2.6.2
click==7.1.1
configparser==4.0.2
cryptography==2.8
decorator==4.0.11
EasyProcess==0.2.10
emoji==0.5.4
Flask==1.1.1
future==0.18.2
googleapis-common-protos==1.51.0
grpcio==1.27.2
grpcio-tools==1.27.2
idna==2.9
imageio==2.1.2
instagram-private-api==1.6.0
InstagramAPI==1.0.2
instapy==0.6.8
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==2.6.0
MarkupSafe==1.1.1
MeaningCloud-python==1.2.2
moviepy==0.2.3.2
numpy==1.18.2
Pillow==7.0.0
plyer==1.4.2
protobuf==3.11.3
pycparser==2.20
pymongo==3.10.1
python-telegram-bot==12.4.2
PyVirtualDisplay==0.2.5
PyYAML==5.3.1
regex==2020.2.20
requests==2.11.1
requests-toolbelt==0.7.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.11.2
urllib3==1.25.8
webdriverdownloader==1.1.0.3
Werkzeug==1.0.0
youtube-dl==2020.3.8
